pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "foveaenv"
version = "1.0.0"
description = "Privacy-aware attention navigation benchmark"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "numpy>=1.26.4",
    "gymnasium==0.28.1",
    "stable-baselines3==2.0.0",
    "torch==2.1.0",
    "openai==1.30.0",
    "python-dotenv==1.1.1",
    "requests==2.32.3",
    "PyYAML==6.0.1",
    "openenv-core>=0.2.0",
]

[project.scripts]
server = "app:main"

[tool.setuptools]
py-modules = ["app", "server", "env", "models", "grader"]